Conflicts:
	Corale.Colore.sln
	Corale.Colore.sln.DotSettings

+++ Corale.Colore.Tests/Razer/Keyboard/Effects/CustomTests.cs
@@ -58,7 +58,7 @@ namespace Corale.Colore.Tests.Razer.Keyboard.Effects
                       .EqualTo(-1));
 
             Assert.That(
-                () => dummy = grid[(int)Constants.MaxRows + 1, 0],
+                () => dummy = grid[Constants.MaxRows + 1, 0],
                 Throws.InstanceOf<ArgumentOutOfRangeException>()
                       .With.Property("ParamName")
                       .EqualTo("row")
@@ -74,7 +74,7 @@ namespace Corale.Colore.Tests.Razer.Keyboard.Effects
                       .EqualTo(-1));
 
             Assert.That(
-                () => dummy = grid[0, (int)Constants.MaxColumns + 1],
+                () => dummy = grid[0, Constants.MaxColumns + 1],
                 Throws.InstanceOf<ArgumentOutOfRangeException>()
                       .With.Property("ParamName")
                       .EqualTo("column")
@@ -96,7 +96,7 @@ namespace Corale.Colore.Tests.Razer.Keyboard.Effects
                       .EqualTo(-1));
 
             Assert.That(
-                () => grid[(int)Constants.MaxRows + 1, 0] = Color.Red,
+                () => grid[Constants.MaxRows + 1, 0] = Color.Red,
                 Throws.InstanceOf<ArgumentOutOfRangeException>()
                       .With.Property("ParamName")
                       .EqualTo("row")
@@ -112,7 +112,7 @@ namespace Corale.Colore.Tests.Razer.Keyboard.Effects
                       .EqualTo(-1));
 
             Assert.That(
-                () => grid[0, (int)Constants.MaxColumns + 1] = Color.Red,
+                () => grid[0, Constants.MaxColumns + 1] = Color.Red,
                 Throws.InstanceOf<ArgumentOutOfRangeException>()
                       .With.Property("ParamName")
                       .EqualTo("column")
